Bugfixes, Change structure of tests, Added Testcase
