+++ Paintroid/src/androidTest/java/org/catrobat/paintroid/test/espresso/intro/PossibilitiesTapTargetTest.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.catrobat.paintroid.test.espresso.ui;
+package org.catrobat.paintroid.test.espresso.intro;
 
 import android.support.test.filters.LargeTest;
 import android.support.test.runner.AndroidJUnit4;
@@ -25,7 +25,7 @@ import android.support.test.runner.AndroidJUnit4;
 import org.catrobat.paintroid.R;
 import org.catrobat.paintroid.test.espresso.util.EspressoUtils;
 import org.catrobat.paintroid.test.espresso.util.IntroUtils;
-import org.catrobat.paintroid.test.espresso.util.base.IntroTestBase;
+import org.catrobat.paintroid.test.espresso.intro.base.IntroTestBase;
 import org.catrobat.paintroid.tools.ToolType;
 import org.junit.Before;
 import org.junit.Test;
@@ -39,70 +39,30 @@ import static org.hamcrest.Matchers.not;
 
 @RunWith(AndroidJUnit4.class)
 @LargeTest
-public class IntroToolsTapTargetTest extends IntroTestBase {
+public class PossibilitiesTapTargetTest extends IntroTestBase {
 
     @Before
     public void setUp() throws NoSuchFieldException, IllegalAccessException {
-        introSlide = IntroUtils.IntroSlide.Tools;
+        introSlide = IntroUtils.IntroSlide.Possibilities;
+        startSequence = false;
         super.setUpAndLaunchActivity();
-        EspressoUtils.changeIntroPage(getPageIndexFormLayout(R.layout.islide_tools));
-        EspressoUtils.waitMillis(100);
+        EspressoUtils.changeIntroPage(getPageIndexFormLayout(R.layout.islide_possibilities));
     }
 
     @Test
-    public void testToolBrush() {
-        IntroUtils.introClickToolAndCheckView(ToolType.BRUSH, introSlide);
+    public void testWithoutSequenceToolUndo() throws NoSuchFieldException, IllegalAccessException {
+        IntroUtils.introClickToolAndCheckView(ToolType.UNDO, introSlide);
     }
-
     @Test
-    public void testToolShape() {
-        IntroUtils.introClickToolAndCheckView(ToolType.SHAPE, introSlide);
+    public void testWithoutSequenceToolRedo() throws NoSuchFieldException, IllegalAccessException {
+        IntroUtils.introClickToolAndCheckView(ToolType.REDO, introSlide);
     }
-
     @Test
-    public void testToolTransform() {
-        IntroUtils.introClickToolAndCheckView(ToolType.TRANSFORM, introSlide);
+    public void testWithoutSequenceToolLayer() throws NoSuchFieldException, IllegalAccessException {
+        IntroUtils.introClickToolAndCheckView(ToolType.LAYER, introSlide);
     }
-
     @Test
-    public void testToolLine() {
-        IntroUtils.introClickToolAndCheckView(ToolType.LINE, introSlide);
+    public void testWithoutSequenceToolColorchooser() throws NoSuchFieldException, IllegalAccessException {
+        IntroUtils.introClickToolAndCheckView(ToolType.COLORCHOOSER, introSlide);
     }
-
-    @Test
-    public void testToolCursor() {
-        IntroUtils.introClickToolAndCheckView(ToolType.CURSOR, introSlide);
-    }
-
-    @Test
-    public void testToolFill() {
-        IntroUtils.introClickToolAndCheckView(ToolType.FILL, introSlide);
-    }
-
-    @Test
-    public void testToolPipette() {
-        IntroUtils.introClickToolAndCheckView(ToolType.PIPETTE, introSlide);
-    }
-
-    @Test
-    public void testToolStamp() {
-        IntroUtils.introClickToolAndCheckView(ToolType.STAMP, introSlide);
-    }
-
-    @Test
-    public void testToolImport() {
-        IntroUtils.introClickToolAndCheckView(ToolType.IMPORTPNG, introSlide);
-    }
-
-
-    @Test
-    public void testToolErase() {
-        IntroUtils.introClickToolAndCheckView(ToolType.ERASER, introSlide);
-    }
-
-    @Test
-    public void testToolText() {
-        IntroUtils.introClickToolAndCheckView(ToolType.TEXT, introSlide);
-    }
-
 }
