+++ Paintroid/src/main/java/org/catrobat/paintroid/Session.java
@@ -46,7 +46,7 @@ public class Session {
 	}
 
 	public boolean isFirstTimeLaunch() {
-		return pref.getBoolean(IS_FIRST_TIME_LAUNCH, false);
+		return pref.getBoolean(IS_FIRST_TIME_LAUNCH, true);
 	}
 
 }
